requirements.txt
together>=0.1.0
openai>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.1.0
numpy>=1.24.0
tenacity>=8.2.3
python-dotenv>=1.0.0
pyyaml>=6.0.1
typing-extensions>=4.8.0




